core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5122 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/persist/TableSpaceManagerDefault.java
@@ -100,7 +100,7 @@ public class TableSpaceManagerDefault implements TableSpaceManager {
         }
 
         if (pos < Integer.MAX_VALUE) {
-            lookup.add((int) pos, rowSize);
+            lookup.add(pos, rowSize);
 
             freeBlockSize += rowSize;
         }
@@ -204,7 +204,7 @@ public class TableSpaceManagerDefault implements TableSpaceManager {
             freeBlockSize -= rowSize;
 
             if (difference >= midSize) {
-                int pos = key + ((int) rowSize / scale);
+                long pos = key + (rowSize / scale);
 
                 lookup.add(pos, difference);
             } else {
